backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0
google-generativeai==0.3.0
requests==2.31.0
Pillow==10.1.0
gunicorn==21.2.0
pandas==2.3.3
openpyxl==3.1.5
